# app/handlers/sale.go
package handlers

import (
	"context"
	"encoding/json"
	"io"
	"net/http"
	"strconv"
	"tourmanager/config"
	"tourmanager/core/models"
	"tourmanager/core/ports"
	"tourmanager/util"

	"github.com/gin-gonic/gin"
)

// SetUserRoutes creates sale routes
func SetSaleRoutes(ctx context.Context, cfg config.Config, r *gin.Engine, p ports.SaleService) {
	//	r.Use(middlewares.Recover())

	r.POST("/api/v3.5/sale", createSale(ctx, cfg, p))
	r.GET("/api/v3.5/sale", getAllSale(ctx, cfg, p))
	r.GET("/api/v3.5/sale/page/:page", getAllSalePage(ctx, cfg, p))
	r.GET("/api/v3.5/sale/informe", getInfSale(ctx, cfg, p))
	r.GET("/api/v3.5/sale/:id", getSaleByID(ctx, cfg, p))
	r.PATCH("/api/v3.5/sale/:id", updateSale(ctx, cfg, p))
	r.DELETE("/api/v3.5/sale/:id", deleteSale(ctx, cfg, p))
}

// @Summary Create sale
// @Description Creates a new sale
// @Tags sale
// @Param user body models.CreateSaleReq true "New sale to be created"
// @Success 201 {object} models.SaleResp "OK"
// @Failure 400 {object} object
// @Failure 408 {object} object
// @Failure 500 {object} object
// @Security Bearer
// @Router /api/v3.5/sale [post]
func createSale(ctx context.Context, cfg config.Config, p ports.SaleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		// El ctx ya tiene timeout + schema
		ctx := c.Request.Context()

		// Leer el cuerpo de la solicitud
		body, err := io.ReadAll(c.Request.Body)
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}

		// Parsear el JSON al modelo adecuado
		var sale models.CreateSaleReq
		err = json.Unmarshal(body, &sale)
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}

		// Llamar al servicio para crear los sale
		id, err := p.Create(ctx, sale)
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}

		// Devolver la respuesta con el código de estado adecuado
		result := util.NewMessageResponse("Registro Creado con id: "+id, id)

		response := util.NewSuccessResponse(result, http.StatusAccepted)
		c.JSON(response.StatusCode, response)
	}
}

// @Summary Get inf all ingreso
// @Description Gets all the ingreso
// @Tags ingreso
// @Success 200 {array} models.SaleResp "OK"
// @Failure 400 {object} object
// @Failure 401 {object} object
// @Failure 408 {object} object
// @Failure 500 {object} object
// @Security Bearer
// @Router /api/v3.5/ingreso [get]
func getInfSale(ctx context.Context, cfg config.Config, p ports.SaleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		// El ctx ya tiene timeout + schema
		ctx := c.Request.Context()

		filter := make(map[string]interface{})

		// Obtener todos los parámetros de consulta
		queryParams := c.Request.URL.Query()
		// Convertir a un map[string]interface{}
		for key, values := range queryParams {
			if len(values) > 1 {
				// Si hay múltiples valores, los almacenamos como slice
				filter[key] = values
			} else {
				// Si hay un solo valor, lo almacenamos directamente
				filter[key] = values[0]
			}
		}

		result, err := p.GetInforme(ctx, filter)
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}
		// Agregar el total count al header (fuera del wrapper JSON)
		c.Header("X-Total-Count", strconv.FormatInt(result.TotalCount, 10))

		// Enviar solo los items en el wrapper de éxito
		response := util.NewSuccessResponse(result.Items, http.StatusOK)
		c.JSON(response.StatusCode, response)
	}

}

// @Summary Get all sale
// @Description Gets all the sale
// @Tags sale
// @Success 200 {array} models.SaleResp "OK"
// @Failure 400 {object} object
// @Failure 401 {object} object
// @Failure 408 {object} object
// @Failure 500 {object} object
// @Security Bearer
// @Router /api/v3.5/sale [get]
func getAllSale(ctx context.Context, cfg config.Config, p ports.SaleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		// El ctx ya tiene timeout + schema
		ctx := c.Request.Context()

		filter := make(map[string]interface{})

		// Obtener todos los parámetros de consulta
		queryParams := c.Request.URL.Query()
		// Convertir a un map[string]interface{}
		for key, values := range queryParams {
			if len(values) > 1 {
				// Si hay múltiples valores, los almacenamos como slice
				filter[key] = values
			} else {
				// Si hay un solo valor, lo almacenamos directamente
				filter[key] = values[0]
			}
		}

		result, err := p.GetAll(ctx, filter)
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}
		// Agregar el total count al header (fuera del wrapper JSON)
		c.Header("X-Total-Count", strconv.FormatInt(result.TotalCount, 10))

		// Enviar solo los items en el wrapper de éxito
		response := util.NewSuccessResponse(result.Items, http.StatusOK)
		c.JSON(response.StatusCode, response)
	}
}

// @Summary Get one sale by page
// @Description Gets one sale by page
// @Tags sale
// @Success 200 {array} models.SaleResp "OK"
// @Failure 400 {object} object
// @Failure 401 {object} object
// @Failure 408 {object} object
// @Failure 500 {object} object
// @Security Bearer
// @Router /api/v3.5/sale/page/:page [get]
func getAllSalePage(ctx context.Context, cfg config.Config, p ports.SaleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		// El ctx ya tiene timeout + schema
		ctx := c.Request.Context()

		page, err := strconv.Atoi(c.Param("page"))
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}

		result, err := p.GetAllPage(ctx, page)
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}

		// Enviar solo los items en el wrapper de éxito
		response := util.NewSuccessResponsePage(result.TotalCount, result.Page, result.PageSize, result.TotalPages, http.StatusOK, result.Items)
		c.JSON(response.StatusCode, response)
	}
}

// @Summary Get sale by ID
// @Description Gets a sale by ID
// @Tags sale
// @Param id path string true "ID"
// @Success 200 {object} models.SaleResp "OK"
// @Failure 400 {object} object
// @Failure 401 {object} object
// @Failure 408 {object} object
// @Failure 500 {object} object
// @Security Bearer
// @Router /api/v3.5/sale/{id} [get]
func getSaleByID(ctx context.Context, cfg config.Config, p ports.SaleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		// El ctx ya tiene timeout + schema
		ctx := c.Request.Context()

		//	var params = mux.Vars(c.Request)
		result, err := p.GetByID(ctx, c.Param("id"))
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}
		//	utils.ResponseJSON(c.Writer, c.Request, nil, http.StatusAccepted, users)
		response := util.NewSuccessResponse(result, http.StatusOK)
		c.JSON(response.StatusCode, response)
	}
}

// @Summary Update sale
// @Description Updates a sale
// @Tags sale
// @Param id path string true "ID"
// @Param User body models.UpdateSaleReq true "sale"
// @Success 200 "OK"
// @Failure 400 {object} object
// @Failure 401 {object} object
// @Failure 408 {object} object
// @Failure 500 {object} object
// @Security Bearer
// @Router /api/v3.5/users/{id} [patch]
func updateSale(ctx context.Context, cfg config.Config, p ports.SaleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		// El ctx ya tiene timeout + schema
		ctx := c.Request.Context()

		body, err := io.ReadAll(c.Request.Body)
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}

		params := c.Params
		var sale models.UpdateSaleReq
		err = json.Unmarshal(body, &sale)
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}
		id := params.ByName("id")
		err = p.Update(ctx, params.ByName("id"), sale)
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}
		result := util.NewMessageResponse("Registro con id "+id+" actualizado", id)

		response := util.NewSuccessResponse(result, http.StatusAccepted)
		c.JSON(response.StatusCode, response)
	}
}

// @Summary Delete sale
// @Description Delete a sale
// @Tags sale
// @Param id path string true "ID"
// @Success 200 "OK"
// @Failure 400 {object} object
// @Failure 401 {object} object
// @Failure 408 {object} object
// @Failure 500 {object} object
// @Security Bearer
// @Router /api/v3.5/sale/{id} [delete]
func deleteSale(ctx context.Context, cfg config.Config, p ports.SaleService) gin.HandlerFunc {
	return func(c *gin.Context) {
		// El ctx ya tiene timeout + schema
		ctx := c.Request.Context()

		//	var params = mux.Vars(c.Request)
		id := c.Param("id")
		err := p.Delete(ctx, c.Param("id"))
		if err != nil {
			response := util.NewErrorResponse(err, http.StatusInternalServerError)
			c.JSON(response.StatusCode, response)
			return
		}
		result := util.NewMessageResponse("Registro con id "+id+" eliminado ", id)

		response := util.NewSuccessResponse(result, http.StatusAccepted)
		c.JSON(response.StatusCode, response)
	}
}